move PoolFactory to v1
